move paths to their own file
