+++ __config__/webpack/paths.js
@@ -1,0 +1,4 @@
+// @flow
+const path = require('path');
+
+module.exports.dist = path.resolve(__dirname, '..', '..', 'dist');
